Redirect to https onload
This redirect the github.io site onload to https.
This does not redirect localhost or 127.0.0.1 to htpps (praise the dev).

Caution : this will make informations stored on the http://scrumble.github.io localstorage unavalable
So far I only found that it marks as unread the "What's new in scrumble" content.